subt: move simulation from test_subt.py
